requirements.txt
langchain
python-dotenv
langchain_groq
langchain_google_genai
langchain_community
langchain-openai
pypdf
faiss-cpu
structlog
PyMuPDF
pandas
-e .